Refactor: 코드 중복 처리, ViewHolder 분리, 패키지 추가
View색상만 다르기에 작동하는부분에서 색상을 바꿔주고 View를 통일시켰습니다.
ViewHolder를 분리시키고 adpter관련 패키지를 만들어 이동하였습니다.
