+++ PlanJ/app/src/main/java/com/boostcamp/planj/ui/main/MyFragmentPagerAdapter.kt
@@ -1,8 +1,8 @@
 package com.boostcamp.planj.ui.main
 
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.boostcamp.planj.ui.main.week.WeekFragment
 
 class MyFragmentPagerAdapter(fragment : Fragment) : FragmentStateAdapter(fragment) {
