+++ PlanJ/app/src/main/java/com/boostcamp/planj/ui/main/week/adapter/TodayScheduleViewHolder.kt
@@ -1,0 +1,16 @@
+package com.boostcamp.planj.ui.main.week.adapter
+
+import androidx.recyclerview.widget.RecyclerView
+import com.boostcamp.planj.data.model.Schedule
+import com.boostcamp.planj.databinding.ItemWeekScheduleBinding
+
+
+class TodayScheduleViewHolder(binding: ItemWeekScheduleBinding) :
+    RecyclerView.ViewHolder(binding.root) {
+
+    private val title = binding.vwScheduleTitle
+
+    fun bind(item: Schedule) {
+        title.text = item.title
+    }
+}
