--- PlanJ/app/src/main/res/layout/item_week_start.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="1dp">
-
-    <View
-        android:id="@+id/view_weekend_bar"
-        android:layout_width="match_parent"
-        android:layout_height="7dp"
-        android:layout_margin="2dp"
-        android:background="@color/green"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/tv_week_start_title"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="4dp"
-        android:ellipsize="end"
-        android:maxLines="1"
-        android:text="가가가각가가가"
-        android:textSize="10sp"
-        app:layout_constraintEnd_toEndOf="@id/view_weekend_bar"
-        app:layout_constraintStart_toStartOf="@id/view_weekend_bar"
-        app:layout_constraintTop_toBottomOf="@id/view_weekend_bar" />
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="0.5dp"
-        android:layout_marginTop="7dp"
-        android:background="@color/black"
-        app:layout_constraintEnd_toEndOf="@id/view_weekend_bar"
-        app:layout_constraintStart_toStartOf="@id/view_weekend_bar"
-        app:layout_constraintTop_toBottomOf="@+id/tv_week_start_title" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
